[refactor] read / write 코드 스타일 수정 read / write unittest code 통합 컴파일 오류 수정
