+++ virtualSSD-Test/virtualSSD-Test.vcxproj
@@ -99,8 +99,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\packages\gmock.1.11.0\lib\native\src\gtest\src\gtest_main.cc" />
-    <ClCompile Include="SSD_Read_Unittest.cpp" />
-    <ClCompile Include="SSD_Write_Unittest.cpp" />
+    <ClCompile Include="SSD_Unittest.cpp" />
     <ClCompile Include="test.cpp" />
   </ItemGroup>
   <ItemGroup>
