+++ virtual_SSD/SSD.cpp
@@ -8,11 +8,11 @@
 #include <map>
 #include"SSD.h"
 
-class LBARangeException :public std::exception {
-};
+class LBARangeException : public std::exception {};
 
-class DataRangeException :public std::exception {
-};
+class DataRangeException : public std::exception {};
+
+class NotExistNandFileException : public std::exception {};
 
 void SSD::Write(const int& LBA, const std::string& data) {
     CheckWriteCondition(LBA, data);
